Feat: Add BorderlessTextField to MessageForm
