--- src/components/BorderlessTextField.style.ts
@@ -1,10 +1,0 @@
-export const removeBorder = {
-  '& .MuiOutlinedInput-notchedOutline': {
-    border: 'none',
-  },
-  '&.Mui-focused': {
-    '& .MuiOutlinedInput-notchedOutline': {
-      border: 'none',
-    },
-  },
-}
